refactor: reuse minimal index alias metadata
