+++ src/Foundatio.Repositories.Elasticsearch/Extensions/ElasticIndexExtensions.cs
@@ -25,15 +25,15 @@ public static class ElasticIndexExtensions
 {
     /// <summary>
     /// Builds a <see cref="GetIndexRequest"/> that resolves the names of the indexes matching
-    /// <paramref name="indices"/> without materializing their mappings or settings.
+    /// <paramref name="indices"/> without materializing their mapping properties or settings content.
     /// </summary>
     /// <param name="indices">The index name(s) or pattern(s) to resolve (e.g. a wildcard or <c>Indices.All</c>).</param>
     /// <param name="ignoreUnavailable">When <see langword="true"/>, missing or closed indexes are ignored instead of producing an error.</param>
     /// <remarks>
     /// Requesting only the <see cref="Feature.Aliases"/> feature (together with
-    /// <c>IncludeDefaults = false</c>) tells Elasticsearch to omit the large <c>mappings</c> and
-    /// <c>settings</c> sections from the response. This avoids the substantial memory amplification
-    /// the typed client incurs while deserializing mappings for every matching index, which can
+    /// <c>IncludeDefaults = false</c>) strips the large <c>mappings</c> and <c>settings</c> content from
+    /// the response. This avoids the substantial memory amplification the typed client incurs while
+    /// deserializing mapping properties for every matching index, which can
     /// throw an <see cref="OutOfMemoryException"/> when a pattern matches many indexes that each have
     /// large mappings. The response still exposes every matching index name via its <c>Indices</c>
     /// keys. See https://github.com/elastic/elasticsearch-net/issues/8919.
