+++ src/Foundatio.Repositories.Elasticsearch/Jobs/CleanupIndexesJob.cs
@@ -57,9 +57,6 @@ public class CleanupIndexesJob : IJob
 
         var sw = Stopwatch.StartNew();
         // Note: ResolveIndexAsync sends a body in ES 9.x client which ES rejects; use GetAsync instead.
-        // Request only the index names (no mappings/settings) so listing every index in the cluster does
-        // not materialize their mappings. See ElasticIndexExtensions.CreateGetIndexNamesRequest /
-        // https://github.com/elastic/elasticsearch-net/issues/8919.
         var request = ElasticIndexExtensions.CreateGetIndexNamesRequest(Indices.All, ignoreUnavailable: true);
         request.RequestConfiguration = new RequestConfiguration { RequestTimeout = TimeSpan.FromMinutes(5) };
         var result = await _client.Indices.GetAsync(request, cancellationToken).AnyContext();
